+++ CHANGELOG.md
@@ -7,6 +7,7 @@ omnibus-gitlab repository.
 
 - Include OSS driver for object storage support in the registry
 - Support setting of PostgreSQL SSL mode and compression in repmgr conninfo
+- Disable PostgreSQL SSL compression by default
 - Tighten permission on Mattermost's config.json file !2587
 - Tighten permission on Gitaly's config.toml file !2589
 - Tighten permission on gitlab-shell's config.yml file !2585